[AF021]  Created the second act for the intent
Secound activity have been created that is started with an intent in the openNewActivityWithNewText method. For every different button click in the goToText method the method openNewActivityWithNewText with a different text is called.
The text is fetched from the strings.xml with the right id. In the ActText1 activity the intent is fetched with getIntetn() and the extra is extracted from it. With getStringExtra and the key final static String EXTRA_TEXT from MainActivity. The text is then passed to the text view in the Scroll View of ActText1.